+++ src/mjlab/asset_zoo/robots/unitree_d1/__init__.py
@@ -1,0 +1,1 @@
+"""Unitree D1 arm asset."""