+++ src/mjlab/asset_zoo/robots/unitree_go2_d1/__init__.py
@@ -1,0 +1,1 @@
+"""Unitree Go2 with mounted D1 arm asset."""